+++ test/test_auto_unchoke.cpp
@@ -67,7 +67,7 @@ void test_swarm()
 	pack.set_int(settings_pack::alert_mask, alert::all_categories);
 	pack.set_bool(settings_pack::allow_multiple_connections_per_ip, true);
 	pack.set_int(settings_pack::choking_algorithm, settings_pack::rate_based_choker);
-	pack.set_int(settings_pack::upload_rate_limit, rate_limit);
+	pack.set_int(settings_pack::upload_rate_limit, int(rate_limit));
 	pack.set_int(settings_pack::unchoke_slots_limit, 1);
 	pack.set_int(settings_pack::max_retry_port_bind, 900);
 	pack.set_str(settings_pack::listen_interfaces, "0.0.0.0:48010");
@@ -80,8 +80,8 @@ void test_swarm()
 
 	lt::session ses1(pack);
 
-	pack.set_int(settings_pack::upload_rate_limit, rate_limit / 10);
-	pack.set_int(settings_pack::download_rate_limit, rate_limit / 5);
+	pack.set_int(settings_pack::upload_rate_limit, int(rate_limit / 10));
+	pack.set_int(settings_pack::download_rate_limit, int(rate_limit / 5));
 	pack.set_int(settings_pack::unchoke_slots_limit, 0);
 	pack.set_int(settings_pack::choking_algorithm, settings_pack::fixed_slots_choker);
 	pack.set_str(settings_pack::listen_interfaces, "0.0.0.0:49010");
